Move Multer helper functions to a helper folder
