+++ helpers.js
@@ -1,12 +1,27 @@
 /* eslint-disable func-names */
+// Como la aplicacion es pequeña, todos los route handlers estan en un mismo archivo
+const multer = require('multer');
+const path = require('path');
+
 const imageFilter = function (req, file, cb) {
   // Accept images only
+  console.log('entra al image filter');
   if (!file.originalname.match(/\.(jpg|JPG|jpeg|JPEG|png|PNG|gif|GIF)$/)) {
     req.fileValidationError = 'Only image files are allowed!';
     return cb(new Error('Only image files are allowed!'), false);
   }
   cb(null, true);
 };
-exports.imageFilter = imageFilter;
 
-// TODO ponerlo en una carpeta propia
+const storage = multer.diskStorage({
+  destination(req, file, cb) {
+    cb(null, 'uploads/');
+  },
+
+  // By default, multer removes file extensions this will add them back
+  filename(req, file, cb) {
+    cb(null, `${file.fieldname}-${Date.now()}${path.extname(file.originalname)}`);
+  },
+});
+
+module.exports = { imageFilter, storage };
